# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Onfly API",
        "description": "API de gestão de pedidos Onfly.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Ambiente local"
        }
    ],
    "paths": {
        "/auth/token": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Gerar token de acesso JWT",
                "description": "Autentica o usuário via x_api_key e retorna um token JWT.",
                "operationId": "d533fc0dd1c51d7f9b2c7504bf5b35d7",
                "parameters": [
                    {
                        "name": "X-Request-Id",
                        "in": "header",
                        "description": "Identificador único da requisição (UUID).",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "x_api_key"
                                ],
                                "properties": {
                                    "x_api_key": {
                                        "description": "Chave de API do usuário.",
                                        "type": "string",
                                        "example": "abc123xyz..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token gerado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespostaToken"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Header X-Request-Id ausente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespostaErro"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credencial inválida.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespostaErro"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados de entrada inválidos.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespostaErro"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "RespostaErro": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Mensagem de erro."
                    }
                },
                "type": "object"
            },
            "RespostaToken": {
                "properties": {
                    "data": {
                        "properties": {
                            "access_token": {
                                "type": "string",
                                "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                            },
                            "token_type": {
                                "type": "string",
                                "example": "bearer"
                            },
                            "expires_in": {
                                "type": "string",
                                "example": "3600"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT obtido em POST /auth/token.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}